+++ src/main/java/ru/zhendozzz/vkbot/dao/repository/JobLogRepository.java
@@ -12,6 +12,6 @@ import ru.zhendozzz.vkbot.dao.entity.JobLog;
 
 @Repository
 public interface JobLogRepository extends CrudRepository<JobLog, Long> {
-    Optional<JobLog> findByGroupIdAndDateAndSuccessAndType(Integer groupId, LocalDate date, Boolean success, String type);
+    Optional<JobLog> findByGroupIdAndDateAndSuccessAndType(Long groupId, LocalDate date, Boolean success, String type);
     List<JobLog> findAll();
 }
